+++ src/scanoss/api/components/v2/scanoss_components_pb2_grpc.py
@@ -32,6 +32,11 @@ class ComponentsStub(object):
                 request_serializer=scanoss_dot_api_dot_components_dot_v2_dot_scanoss__components__pb2.CompVersionRequest.SerializeToString,
                 response_deserializer=scanoss_dot_api_dot_components_dot_v2_dot_scanoss__components__pb2.CompVersionResponse.FromString,
                 )
+        self.GetComponentStatistics = channel.unary_unary(
+                '/scanoss.api.components.v2.Components/GetComponentStatistics',
+                request_serializer=scanoss_dot_api_dot_common_dot_v2_dot_scanoss__common__pb2.PurlRequest.SerializeToString,
+                response_deserializer=scanoss_dot_api_dot_components_dot_v2_dot_scanoss__components__pb2.CompStatisticResponse.FromString,
+                )
 
 
 class ComponentsServicer(object):
@@ -41,10 +46,6 @@ class ComponentsServicer(object):
 
     def Echo(self, request, context):
         """Standard echo
-        option (google.api.http) = {
-        post: "/api/v2/components/echo"
-        body: "*"
-        };
         """
         context.set_code(grpc.StatusCode.UNIMPLEMENTED)
         context.set_details('Method not implemented!')
@@ -52,10 +53,6 @@ class ComponentsServicer(object):
 
     def SearchComponents(self, request, context):
         """Search for components
-        option (google.api.http) = {
-        post: "/api/v2/components/search"
-        body: "*"
-        };
         """
         context.set_code(grpc.StatusCode.UNIMPLEMENTED)
         context.set_details('Method not implemented!')
@@ -63,10 +60,13 @@ class ComponentsServicer(object):
 
     def GetComponentVersions(self, request, context):
         """Get all version information for a specific component
-        option (google.api.http) = {
-        post: "/api/v2/components/versions"
-        body: "*"
-        };
+        """
+        context.set_code(grpc.StatusCode.UNIMPLEMENTED)
+        context.set_details('Method not implemented!')
+        raise NotImplementedError('Method not implemented!')
+
+    def GetComponentStatistics(self, request, context):
+        """Get the statistics for the specified components
         """
         context.set_code(grpc.StatusCode.UNIMPLEMENTED)
         context.set_details('Method not implemented!')
@@ -89,6 +89,11 @@ def add_ComponentsServicer_to_server(servicer, server):
                     servicer.GetComponentVersions,
                     request_deserializer=scanoss_dot_api_dot_components_dot_v2_dot_scanoss__components__pb2.CompVersionRequest.FromString,
                     response_serializer=scanoss_dot_api_dot_components_dot_v2_dot_scanoss__components__pb2.CompVersionResponse.SerializeToString,
+            ),
+            'GetComponentStatistics': grpc.unary_unary_rpc_method_handler(
+                    servicer.GetComponentStatistics,
+                    request_deserializer=scanoss_dot_api_dot_common_dot_v2_dot_scanoss__common__pb2.PurlRequest.FromString,
+                    response_serializer=scanoss_dot_api_dot_components_dot_v2_dot_scanoss__components__pb2.CompStatisticResponse.SerializeToString,
             ),
     }
     generic_handler = grpc.method_handlers_generic_handler(
@@ -150,5 +155,22 @@ class Components(object):
         return grpc.experimental.unary_unary(request, target, '/scanoss.api.components.v2.Components/GetComponentVersions',
             scanoss_dot_api_dot_components_dot_v2_dot_scanoss__components__pb2.CompVersionRequest.SerializeToString,
             scanoss_dot_api_dot_components_dot_v2_dot_scanoss__components__pb2.CompVersionResponse.FromString,
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
+
+    @staticmethod
+    def GetComponentStatistics(request,
+            target,
+            options=(),
+            channel_credentials=None,
+            call_credentials=None,
+            insecure=False,
+            compression=None,
+            wait_for_ready=None,
+            timeout=None,
+            metadata=None):
+        return grpc.experimental.unary_unary(request, target, '/scanoss.api.components.v2.Components/GetComponentStatistics',
+            scanoss_dot_api_dot_common_dot_v2_dot_scanoss__common__pb2.PurlRequest.SerializeToString,
+            scanoss_dot_api_dot_components_dot_v2_dot_scanoss__components__pb2.CompStatisticResponse.FromString,
             options, channel_credentials,
             insecure, call_credentials, compression, wait_for_ready, timeout, metadata)